refactor: reorganize file, add exception handler, add constant
